+++ app/models/DemarcheComponent.php
@@ -186,4 +186,7 @@ abstract class DemarcheComponent extends RevisableModel {
 		}
 		return $query;
 	}
+	public function scopeExpertisesIds($query, $ids) {
+		return $query;
+	}
 }