+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Dashboard
 - Added :  Ajout du filtre "par action" dans la fonction "Mes filtres"
+- Added :  Application du filtre "par action" sur les projets de simplif'
 - Changed : Désactivation du filtre "par tag"
 - Bugfix : Les filtres ne s'appliquaient pas sur les pièces et tâches dont le total est présenté dans le pavé "Catalogue des démarches"